flag 컬럼 변경 및 개인 법인 구분 컬럼 추가
