+++ src/main/java/com/example/backend/dto/CompanyDto.java
@@ -22,5 +22,6 @@ public class CompanyDto {
 	private String companyFax;
 	private String companyZipCode;
 	private String companyForeigner;
-	private int flag;
+	private String useYN;
+	private String pcBuisness;
 }
