+++ src/main/java/com/example/backend/service/CompanyService.java
@@ -1,5 +1,15 @@
+
 package com.example.backend.service;
 
+import java.util.List;
+
+import com.example.backend.dto.CompanyDto;
+
 public interface CompanyService {
- 
+	List <CompanyDto> getCompanyInformation();
+	CompanyDto getCompanyInfo(int companySeq);
+	
+	void insertCompany( CompanyDto dto);
+	void updateCompany(CompanyDto dto);
+	void DeleteCompany(int companySeq);
 }
