--- src/main/java/com/example/backend/service/companyservice.java
@@ -1,15 +1,0 @@
-
-package com.example.backend.service;
-
-import java.util.List;
-
-import com.example.backend.dto.CompanyDto;
-
-public interface CompanyService {
-	List <CompanyDto> getCompanyInformation();
-	CompanyDto getCompanyInfo(int companySeq);
-	
-	void insertCompany( CompanyDto dto);
-	void updateCompany(CompanyDto dto);
-	void DeleteCompany(int companySeq);
-}
